Refactor exceptions under one parent TsundereException class
Children exception classes can store the message to be displayed to the user via getMessage().
Thus we do not need a specific function in UI class for every exception message.

Refactoring exceptions reduces the need to add code for every new exceptions in the classes that catch it, and the UI class.

Let's use inheritance to minimize the code length.